refactor: split built-in node registration modules
Move BUILTIN_NODE_TYPES from src/loom.js into 8 library-specific registration modules under src/nodes/. Each module exports a register[X]Nodes(registry) function that calls registry.registerNodeType() for each node in that category.

Changes:
- Create src/nodes/{core,function,math,logic,text,list,scene,dom}.js with node definitions
- Create src/nodes/index.js with registerBuiltinNodes() that calls all registration functions
- Create src/nodes/helpers.js to re-export helper functions needed by node modules
- Update src/loom.js to import and use registerBuiltinNodes() in createDefaultNodeRegistry()
- Remove BUILTIN_NODE_TYPES object (1089 lines) from src/loom.js
- Add test/builtin-node-registration.test.mjs with 4 test cases
- Update package.json to include new test in test:unit

Runtime behavior is unchanged:
- NODE_TYPES export contains identical node definitions
- DEFAULT_NODE_REGISTRY works identically
- createDefaultNodeRegistry() returns a registry with all built-in nodes
- All existing node names and behavior preserved
- All 118 built-in nodes registered through the same API

This prepares Loomlet for external package support, as built-in nodes now use the same registration API that future packages will use.
