+++ src/loom.js
@@ -1,6 +1,7 @@
 // Loom: ブラウザで動くステートレスなデータフロー実行エンジン
 
 import { createNodeRegistry } from './runtime/node-registry.js';
+import { registerBuiltinNodes } from './nodes/index.js';
 
 export class LoomError extends Error {
   constructor(code, message, details = {}) {
@@ -553,1097 +554,29 @@ function getNodePath() {
   return globalThis.process.getBuiltinModule('path');
 }
 
+// Export helper functions for node modules
+export {
+  isFiniteNumber,
+  coerceFiniteNumber,
+  resolveStateInputValue,
+  sanitizeStateValue,
+  stringifyJsonValue,
+  stringifyTextValue,
+  inspectValue,
+  toArray,
+  collectInputs,
+  unsupportedFunctionValueNode,
+  isLoomletCallable,
+  assertLoomletCallable,
+  isLoomletTruthy,
+  evaluateLegacyFunctionExpr,
+  createLoomletFunction,
+  mapFunctionValueNode,
+  getNodeFs,
+  getNodePath
+};
 
 // ノード型レジストリ
-const BUILTIN_NODE_TYPES = {
-  // Phase 0 ノード
-  clock: {
-    category: 'source',
-    inputs: [],
-    outputs: [{ name: 't', type: 'number', kind: 'behavior' }],
-    params: [],
-    evaluate: (inputs, params, ctx) => ({ t: ctx.time })
-  },
-  'function.literal': {
-    category: 'source',
-    inputs: [],
-    outputs: [{ name: 'out', type: 'function', kind: 'behavior' }],
-    params: [
-      { name: 'params', type: 'array', default: [] },
-      { name: 'body', type: 'any', default: null },
-      { name: 'closureRefs', type: 'object', default: {} }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: createLoomletFunction(params.params || [], params.body, params.closureRefs, ctx) })
-  },
-  'function.call': {
-    category: 'transform',
-    inputs: [
-      { name: 'fn', type: 'function', default: null, kind: 'behavior' },
-      ...Array.from({ length: 8 }, (_, i) => ({ name: `arg${i + 1}`, type: 'any', default: undefined, kind: 'behavior' }))
-    ],
-    outputs: [{ name: 'out', type: 'any', kind: 'behavior' }],
-    params: [],
-    evaluate: (inputs, params, ctx) => {
-      const fn = assertLoomletCallable(inputs.fn, 'function.call');
-      return { out: fn.call(collectInputs(inputs, Array.from({ length: 8 }, (_, i) => `arg${i + 1}`)), ctx) };
-    }
-  },
-  constant: {
-    category: 'source',
-    inputs: [],
-    outputs: [{ name: 'out', type: 'any', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: 0 }],
-    evaluate: (inputs, params, ctx) => ({ out: params.value })
-  },
-  sine: {
-    category: 'transform',
-    inputs: [
-      { name: 't', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'freq', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'amplitude', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'phase', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'offset', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'freq', type: 'number', default: 1 },
-      { name: 'amplitude', type: 'number', default: 1 },
-      { name: 'phase', type: 'number', default: 0 },
-      { name: 'offset', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const t = inputs.t;
-      const freq = inputs.freq;
-      const amplitude = inputs.amplitude;
-      const phase = inputs.phase;
-      const offset = inputs.offset;
-      return { out: Math.sin(t * freq * 2 * Math.PI + phase) * amplitude + offset };
-    }
-  },
-  add: {
-    category: 'transform',
-    commutative: true,
-    inputs: [
-      { name: 'a', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'b', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'a', type: 'number', default: 0 },
-      { name: 'b', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: inputs.a + inputs.b })
-  },
-  multiply: {
-    category: 'transform',
-    commutative: true,
-    inputs: [
-      { name: 'a', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'b', type: 'number', default: 1, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'a', type: 'number', default: 1 },
-      { name: 'b', type: 'number', default: 1 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: inputs.a * inputs.b })
-  },
-
-  // 第1陣: 基本演算系
-  subtract: {
-    category: 'transform',
-    inputs: [
-      { name: 'a', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'b', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'a', type: 'number', default: 0 },
-      { name: 'b', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: inputs.a - inputs.b })
-  },
-  divide: {
-    category: 'transform',
-    inputs: [
-      { name: 'a', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'b', type: 'number', default: 1, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'a', type: 'number', default: 0 },
-      { name: 'b', type: 'number', default: 1 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: inputs.b === 0 ? 0 : inputs.a / inputs.b })
-  },
-  mod: {
-    category: 'transform',
-    inputs: [
-      { name: 'a', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'b', type: 'number', default: 1, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'a', type: 'number', default: 0 },
-      { name: 'b', type: 'number', default: 1 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: inputs.b === 0 ? 0 : ((inputs.a % inputs.b) + inputs.b) % inputs.b })
-  },
-  negate: {
-    category: 'transform',
-    inputs: [
-      { name: 'a', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'a', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: -inputs.a })
-  },
-  abs: {
-    category: 'transform',
-    inputs: [
-      { name: 'a', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'a', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: Math.abs(inputs.a) })
-  },
-
-  // 第2陣: 範囲操作系＋コサイン
-  clamp: {
-    category: 'transform',
-    inputs: [
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'min', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'max', type: 'number', default: 1, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'number', default: 0 },
-      { name: 'min', type: 'number', default: 0 },
-      { name: 'max', type: 'number', default: 1 }
-    ],
-    evaluate: (inputs, params, ctx) => ({
-      out: inputs.min > inputs.max ? inputs.min : Math.max(inputs.min, Math.min(inputs.max, inputs.value))
-    })
-  },
-  lerp: {
-    category: 'transform',
-    inputs: [
-      { name: 'a', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'b', type: 'number', default: 1, kind: 'behavior' },
-      { name: 't', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'a', type: 'number', default: 0 },
-      { name: 'b', type: 'number', default: 1 },
-      { name: 't', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: inputs.a + (inputs.b - inputs.a) * inputs.t })
-  },
-  smoothstep: {
-    category: 'transform',
-    inputs: [
-      { name: 'x', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'edge0', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'edge1', type: 'number', default: 1, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'x', type: 'number', default: 0 },
-      { name: 'edge0', type: 'number', default: 0 },
-      { name: 'edge1', type: 'number', default: 1 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const { x, edge0, edge1 } = inputs;
-      if (edge0 === edge1) {
-        return { out: x < edge0 ? 0 : 1 };
-      }
-      let t = (x - edge0) / (edge1 - edge0);
-      t = Math.max(0, Math.min(1, t));
-      return { out: t * t * (3 - 2 * t) };
-    }
-  },
-  map: {
-    category: 'transform',
-    inputs: [
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'inMin', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'inMax', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'outMin', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'outMax', type: 'number', default: 1, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'number', default: 0 },
-      { name: 'inMin', type: 'number', default: 0 },
-      { name: 'inMax', type: 'number', default: 1 },
-      { name: 'outMin', type: 'number', default: 0 },
-      { name: 'outMax', type: 'number', default: 1 },
-      { name: 'clamp', type: 'boolean', default: false }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const { value, inMin, inMax, outMin, outMax } = inputs;
-      if (inMax === inMin) {
-        return { out: outMin };
-      }
-      let t = (value - inMin) / (inMax - inMin);
-      if (params.clamp === true) {
-        t = Math.max(0, Math.min(1, t));
-      }
-      return { out: outMin + (outMax - outMin) * t };
-    }
-  },
-  cosine: {
-    category: 'transform',
-    inputs: [
-      { name: 't', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'freq', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'amplitude', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'phase', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'offset', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'freq', type: 'number', default: 1 },
-      { name: 'amplitude', type: 'number', default: 1 },
-      { name: 'phase', type: 'number', default: 0 },
-      { name: 'offset', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const t = inputs.t;
-      const freq = inputs.freq;
-      const amplitude = inputs.amplitude;
-      const phase = inputs.phase;
-      const offset = inputs.offset;
-      return { out: Math.cos(t * freq * 2 * Math.PI + phase) * amplitude + offset };
-    }
-  },
-  greaterThan: {
-    category: 'transform',
-    inputs: [
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'threshold', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'number', default: 0 },
-      { name: 'threshold', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: inputs.value > inputs.threshold })
-  },
-  lessThan: {
-    category: 'transform',
-    inputs: [
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'threshold', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'number', default: 0 },
-      { name: 'threshold', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => ({ out: inputs.value < inputs.threshold })
-  },
-  smoothLerp: {
-    category: 'state',
-    inputs: [
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'number', default: 0 },
-      { name: 'rate', type: 'number', default: 5 },
-      { name: 'initial', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const prevOut = sanitizeStateValue(ctx.prevOut, params.initial);
-      const value = resolveStateInputValue(inputs.value, params.initial);
-      const rate = coerceFiniteNumber(params.rate, 5);
-      const factor = 1 - Math.exp(-rate * ctx.dt);
-      return { out: prevOut + (value - prevOut) * factor };
-    }
-  },
-  lowpass: {
-    category: 'state',
-    inputs: [
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'number', default: 0 },
-      { name: 'tau', type: 'number', default: 0.2 },
-      { name: 'initial', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const prevOut = sanitizeStateValue(ctx.prevOut, params.initial);
-      const value = resolveStateInputValue(inputs.value, params.initial);
-      const tau = coerceFiniteNumber(params.tau, 0.2);
-      const factor = tau <= 0 ? 1 : ctx.dt / (tau + ctx.dt);
-      return { out: prevOut + (value - prevOut) * factor };
-    }
-  },
-  delay1: {
-    category: 'state',
-    inputs: [
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'number', default: 0 },
-      { name: 'initial', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const prevOut = sanitizeStateValue(ctx.prevOut, params.initial);
-      const value = resolveStateInputValue(inputs.value, params.initial);
-      return { out: prevOut, _newState: value };
-    }
-  },
-  integrate: {
-    category: 'state',
-    inputs: [
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'number', default: 0 },
-      { name: 'initial', type: 'number', default: 0 },
-      { name: 'min', type: 'number|null', default: null },
-      { name: 'max', type: 'number|null', default: null }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const prevOut = sanitizeStateValue(ctx.prevOut, params.initial);
-      const value = resolveStateInputValue(inputs.value, 0);
-      const min = params.min === null ? null : coerceFiniteNumber(params.min, null);
-      const max = params.max === null ? null : coerceFiniteNumber(params.max, null);
-      let out = prevOut + value * ctx.dt;
-      if (min !== null && out < min) out = min;
-      if (max !== null && out > max) out = max;
-      return { out };
-    }
-  },
-
-  // Phase 1 入力ノード
-  pointerClick: {
-    category: 'input',
-    inputs: [],
-    outputs: [{ name: 'event', type: 'event<vec2>', kind: 'event' }],
-    params: [{ name: 'target', type: 'string', default: 'window' }],
-    evaluate: (inputs, params, ctx) => {
-      // dispatchEvent 経由で this._values に設定済みのため、evaluate は呼ばれない
-      return { event: [] };
-    },
-    onStart: (node, engine) => {
-      const targetSelector = node.params?.target || 'window';
-      const target = targetSelector === 'window' ? window : document.querySelector(targetSelector);
-      if (!target) return;
-
-      const handler = (e) => {
-        engine.dispatchEvent(node.id + '.event', { x: e.clientX, y: e.clientY });
-      };
-      target.addEventListener('pointerdown', handler);
-      node._eventHandler = handler;
-      node._eventTarget = target;
-    },
-    onStop: (node, engine) => {
-      if (node._eventTarget && node._eventHandler) {
-        node._eventTarget.removeEventListener('pointerdown', node._eventHandler);
-        delete node._eventHandler;
-        delete node._eventTarget;
-      }
-    }
-  },
-
-  pointerPosition: {
-    category: 'input',
-    inputs: [],
-    outputs: [{ name: 'pos', type: 'vec2', kind: 'behavior' }],
-    params: [{ name: 'target', type: 'string', default: 'window' }],
-    evaluate: (inputs, params, ctx) => {
-      if (!ctx.engine || !ctx.engine._inputStates) {
-        ctx.engine._inputStates = {};
-      }
-      const lastPos = ctx.engine._inputStates.lastPos || { x: 0, y: 0 };
-      return { pos: lastPos };
-    },
-    onStart: (node, engine) => {
-      const targetSelector = node.params?.target || 'window';
-      const target = targetSelector === 'window' ? window : document.querySelector(targetSelector);
-      if (!target) return;
-
-      if (!engine._inputStates) {
-        engine._inputStates = {};
-      }
-
-      const handler = (e) => {
-        engine._inputStates.lastPos = { x: e.clientX, y: e.clientY };
-      };
-      target.addEventListener('pointermove', handler);
-      node._eventHandler = handler;
-      node._eventTarget = target;
-    },
-    onStop: (node, engine) => {
-      if (node._eventTarget && node._eventHandler) {
-        node._eventTarget.removeEventListener('pointermove', node._eventHandler);
-        delete node._eventHandler;
-        delete node._eventTarget;
-      }
-    }
-  },
-
-  keyDown: {
-    category: 'input',
-    inputs: [],
-    outputs: [{ name: 'event', type: 'event<string>', kind: 'event' }],
-    params: [{ name: 'key', type: 'string', default: null }],
-    evaluate: (inputs, params, ctx) => {
-      // dispatchEvent 経由で this._values に設定済みのため、evaluate は呼ばれない
-      return { event: [] };
-    },
-    onStart: (node, engine) => {
-      const filterKey = node.params?.key || null;
-      const handler = (e) => {
-        if (!filterKey || e.key === filterKey) {
-          engine.dispatchEvent(node.id + '.event', e.key);
-        }
-      };
-      window.addEventListener('keydown', handler);
-      node._eventHandler = handler;
-    },
-    onStop: (node, engine) => {
-      if (node._eventHandler) {
-        window.removeEventListener('keydown', node._eventHandler);
-        delete node._eventHandler;
-      }
-    }
-  },
-
-  keyUp: {
-    category: 'input',
-    inputs: [],
-    outputs: [{ name: 'event', type: 'event<string>', kind: 'event' }],
-    params: [{ name: 'key', type: 'string', default: null }],
-    evaluate: (inputs, params, ctx) => {
-      // dispatchEvent 経由で this._values に設定済みのため、evaluate は呼ばれない
-      return { event: [] };
-    },
-    onStart: (node, engine) => {
-      const filterKey = node.params?.key || null;
-      const handler = (e) => {
-        if (!filterKey || e.key === filterKey) {
-          engine.dispatchEvent(node.id + '.event', e.key);
-        }
-      };
-      window.addEventListener('keyup', handler);
-      node._eventHandler = handler;
-    },
-    onStop: (node, engine) => {
-      if (node._eventHandler) {
-        window.removeEventListener('keyup', node._eventHandler);
-        delete node._eventHandler;
-      }
-    }
-  },
-
-  // Phase 1 イベント変換ノード
-  filter: {
-    category: 'transform',
-    inputs: [{ name: 'event', type: 'event<any>', kind: 'event' }],
-    outputs: [{ name: 'event', type: 'event<any>', kind: 'event' }],
-    params: [{ name: 'predicate', type: 'string', default: 'true' }],
-    evaluate: (inputs, params, ctx) => {
-      const eventPayloads = inputs.event || [];
-      if (!Array.isArray(eventPayloads)) {
-        return { event: [] };
-      }
-
-      // predicate をパース・キャッシュ
-      if (!ctx.nodePredicates) ctx.nodePredicates = new Map();
-      const cacheKey = params.predicate;
-      let evaluator = ctx.nodePredicates.get(cacheKey);
-      if (!evaluator) {
-        try {
-          const dslEval = new RestrictedDSLEvaluator(params.predicate);
-          evaluator = dslEval.evaluate();
-          ctx.nodePredicates.set(cacheKey, evaluator);
-        } catch (e) {
-          throw e; // DSL パースエラーは LoomError で既にラップされている
-        }
-      }
-
-      const filtered = eventPayloads.filter(payload => {
-        try {
-          return evaluator(payload);
-        } catch (e) {
-          return false;
-        }
-      });
-
-      return { event: filtered };
-    }
-  },
-
-  sample: {
-    category: 'transform',
-    inputs: [
-      { name: 'trigger', type: 'event<void>', kind: 'event' },
-      { name: 'value', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'event', type: 'event<number>', kind: 'event' }],
-    params: [],
-    evaluate: (inputs, params, ctx) => {
-      const triggers = inputs.trigger || [];
-      const value = inputs.value;
-
-      if (!Array.isArray(triggers)) {
-        return { event: [] };
-      }
-
-      // trigger が複数回発火した場合、その回数分 value をペイロードに積む
-      const sampled = triggers.map(() => value);
-      return { event: sampled };
-    }
-  },
-
-  merge: {
-    category: 'transform',
-    inputs: [
-      { name: 'a', type: 'event<any>', kind: 'event' },
-      { name: 'b', type: 'event<any>', kind: 'event' }
-    ],
-    outputs: [{ name: 'event', type: 'event<any>', kind: 'event' }],
-    params: [],
-    evaluate: (inputs, params, ctx) => {
-      const aPayloads = inputs.a || [];
-      const bPayloads = inputs.b || [];
-
-      const merged = [
-        ...(Array.isArray(aPayloads) ? aPayloads : []),
-        ...(Array.isArray(bPayloads) ? bPayloads : [])
-      ];
-
-      return { event: merged };
-    }
-  },
-
-  'text.upper': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'string', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }],
-    evaluate: (inputs) => ({ out: String(inputs.value ?? '').toUpperCase() })
-  },
-  'text.lower': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'string', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }],
-    evaluate: (inputs) => ({ out: String(inputs.value ?? '').toLowerCase() })
-  },
-  'text.trim': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'string', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }],
-    evaluate: (inputs) => ({ out: String(inputs.value ?? '').trim() })
-  },
-  'text.replace': {
-    category: 'transform',
-    inputs: [
-      { name: 'value', type: 'any', default: '', kind: 'behavior' },
-      { name: 'search', type: 'any', default: '', kind: 'behavior' },
-      { name: 'replacement', type: 'any', default: '', kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'string', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'any', default: '' },
-      { name: 'search', type: 'any', default: '' },
-      { name: 'replacement', type: 'any', default: '' }
-    ],
-    evaluate: (inputs) => ({
-      out: String(inputs.value ?? '').replaceAll(String(inputs.search ?? ''), String(inputs.replacement ?? ''))
-    })
-  },
-
-  'text.concat': {
-    category: 'transform',
-    inputs: Array.from({ length: 8 }, (_, i) => ({ name: `value${i + 1}`, type: 'any', default: undefined, kind: 'behavior' })),
-    outputs: [{ name: 'out', type: 'string', kind: 'behavior' }],
-    params: Array.from({ length: 8 }, (_, i) => ({ name: `value${i + 1}`, type: 'any', default: undefined })),
-    evaluate: (inputs) => ({ out: collectInputs(inputs, Array.from({ length: 8 }, (_, i) => `value${i + 1}`)).map((value) => stringifyTextValue(value)).join('') })
-  },
-  'text.split': {
-    category: 'transform',
-    inputs: [
-      { name: 'value', type: 'any', default: '', kind: 'behavior' },
-      { name: 'separator', type: 'any', default: ',', kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'array', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }, { name: 'separator', type: 'any', default: ',' }],
-    evaluate: (inputs) => ({ out: stringifyTextValue(inputs.value).split(stringifyTextValue(inputs.separator)) })
-  },
-  'text.join': {
-    category: 'transform',
-    inputs: [
-      { name: 'list', type: 'array', default: [], kind: 'behavior' },
-      { name: 'separator', type: 'any', default: ',', kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'string', kind: 'behavior' }],
-    params: [{ name: 'list', type: 'array', default: [] }, { name: 'separator', type: 'any', default: ',' }],
-    evaluate: (inputs) => ({ out: toArray(inputs.list).map((value) => stringifyTextValue(value)).join(stringifyTextValue(inputs.separator)) })
-  },
-  'text.includes': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }, { name: 'search', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }, { name: 'search', type: 'any', default: '' }],
-    evaluate: (inputs) => ({ out: stringifyTextValue(inputs.value).includes(stringifyTextValue(inputs.search)) })
-  },
-  'text.startsWith': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }, { name: 'search', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }, { name: 'search', type: 'any', default: '' }],
-    evaluate: (inputs) => ({ out: stringifyTextValue(inputs.value).startsWith(stringifyTextValue(inputs.search)) })
-  },
-  'text.endsWith': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }, { name: 'search', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }, { name: 'search', type: 'any', default: '' }],
-    evaluate: (inputs) => ({ out: stringifyTextValue(inputs.value).endsWith(stringifyTextValue(inputs.search)) })
-  },
-  'text.length': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }],
-    evaluate: (inputs) => ({ out: stringifyTextValue(inputs.value).length })
-  },
-  'text.isEmpty': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }],
-    evaluate: (inputs) => ({ out: stringifyTextValue(inputs.value).length === 0 })
-  },
-  'text.stringify': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: null, kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'string', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: null }],
-    evaluate: (inputs) => ({ out: stringifyTextValue(inputs.value) })
-  },
-  'json.parse': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: '', kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'any', kind: 'behavior' }],
-    params: [{ name: 'value', type: 'any', default: '' }],
-    evaluate: (inputs) => {
-      try {
-        return { out: JSON.parse(String(inputs.value ?? '')) };
-      } catch (error) {
-        throw new LoomError('INVALID_JSON', `JSON parse failed: ${error.message}`);
-      }
-    }
-  },
-  'json.stringify': {
-    category: 'transform',
-    inputs: [{ name: 'value', type: 'any', default: null, kind: 'behavior' }],
-    outputs: [{ name: 'out', type: 'string', kind: 'behavior' }],
-    params: [
-      { name: 'value', type: 'any', default: null },
-      { name: 'pretty', type: 'boolean', default: false }
-    ],
-    evaluate: (inputs, params) => ({ out: stringifyJsonValue(inputs.value, params.pretty === true) })
-  },
-  'console.log': {
-    category: 'sink',
-    inputs: [{ name: 'value', type: 'any', default: undefined, kind: 'behavior' }],
-    outputs: [],
-    params: [],
-    evaluate: (inputs, params, ctx) => {
-      ctx.engine?._recordEffect({ type: 'console.log', level: 'log', value: inputs.value, nodeId: ctx.currentNodeId });
-      return {};
-    }
-  },
-  'console.warn': {
-    category: 'sink',
-    inputs: [{ name: 'value', type: 'any', default: undefined, kind: 'behavior' }],
-    outputs: [],
-    params: [],
-    evaluate: (inputs, params, ctx) => {
-      ctx.engine?._recordEffect({ type: 'console.warn', level: 'warn', value: inputs.value, nodeId: ctx.currentNodeId });
-      return {};
-    }
-  },
-  'console.error': {
-    category: 'sink',
-    inputs: [{ name: 'value', type: 'any', default: undefined, kind: 'behavior' }],
-    outputs: [],
-    params: [],
-    evaluate: (inputs, params, ctx) => {
-      ctx.engine?._recordEffect({ type: 'console.error', level: 'error', value: inputs.value, nodeId: ctx.currentNodeId });
-      return {};
-    }
-  },
-
-
-  'console.table': {
-    category: 'sink',
-    inputs: [{ name: 'value', type: 'any', default: undefined, kind: 'behavior' }],
-    outputs: [],
-    params: [],
-    evaluate: (inputs, params, ctx) => {
-      ctx.engine?._recordEffect({ type: 'console.table', level: 'table', value: inputs.value, nodeId: ctx.currentNodeId });
-      if (typeof console.table === 'function' && ctx.emitConsole === true) console.table(inputs.value);
-      return {};
-    }
-  },
-
-  'logic.not': { category: 'transform', inputs: [{ name: 'value', type: 'any', default: false, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'any', default: false }], evaluate: (inputs) => ({ out: !inputs.value }) },
-  'logic.and': { category: 'transform', commutative: true, inputs: [{ name: 'a', type: 'any', default: false, kind: 'behavior' }, { name: 'b', type: 'any', default: false, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'a', type: 'any', default: false }, { name: 'b', type: 'any', default: false }], evaluate: (inputs) => ({ out: Boolean(inputs.a && inputs.b) }) },
-  'logic.or': { category: 'transform', commutative: true, inputs: [{ name: 'a', type: 'any', default: false, kind: 'behavior' }, { name: 'b', type: 'any', default: false, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'a', type: 'any', default: false }, { name: 'b', type: 'any', default: false }], evaluate: (inputs) => ({ out: Boolean(inputs.a || inputs.b) }) },
-  'logic.equals': { category: 'transform', inputs: [{ name: 'value', type: 'any', default: null, kind: 'behavior' }, { name: 'other', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'any', default: null }, { name: 'other', type: 'any', default: null }], evaluate: (inputs) => ({ out: Object.is(inputs.value, inputs.other) }) },
-  'logic.notEquals': { category: 'transform', inputs: [{ name: 'value', type: 'any', default: null, kind: 'behavior' }, { name: 'other', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'any', default: null }, { name: 'other', type: 'any', default: null }], evaluate: (inputs) => ({ out: !Object.is(inputs.value, inputs.other) }) },
-  'logic.greaterThan': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'other', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'other', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.value > inputs.other }) },
-  'logic.lessThan': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'other', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'other', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.value < inputs.other }) },
-  'logic.greaterOrEqual': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'other', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'other', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.value >= inputs.other }) },
-  'logic.lessOrEqual': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'other', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'other', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.value <= inputs.other }) },
-  'logic.select': { category: 'transform', inputs: [{ name: 'condition', type: 'any', default: false, kind: 'behavior' }, { name: 'whenTrue', type: 'any', default: null, kind: 'behavior' }, { name: 'whenFalse', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'condition', type: 'any', default: false }, { name: 'whenTrue', type: 'any', default: null }, { name: 'whenFalse', type: 'any', default: null }], evaluate: (inputs) => ({ out: inputs.condition ? inputs.whenTrue : inputs.whenFalse }) },
-  'logic.when': { category: 'transform', inputs: [{ name: 'condition', type: 'any', default: false, kind: 'behavior' }, { name: 'value', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'condition', type: 'any', default: false }, { name: 'value', type: 'any', default: null }], evaluate: (inputs) => ({ out: inputs.condition ? inputs.value : null }) },
-
-  'list.of': { category: 'transform', inputs: Array.from({ length: 8 }, (_, i) => ({ name: `value${i + 1}`, type: 'any', default: undefined, kind: 'behavior' })), outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: Array.from({ length: 8 }, (_, i) => ({ name: `value${i + 1}`, type: 'any', default: undefined })), evaluate: (inputs) => ({ out: collectInputs(inputs, Array.from({ length: 8 }, (_, i) => `value${i + 1}`)) }) },
-  'list.range': { category: 'transform', inputs: [{ name: 'start', type: 'number', default: 0, kind: 'behavior' }, { name: 'end', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'start', type: 'number', default: 0 }, { name: 'end', type: 'number', default: 0 }], evaluate: (inputs) => { const start = Math.trunc(inputs.start); const end = Math.trunc(inputs.end); const step = start <= end ? 1 : -1; const out = []; for (let n = start; step > 0 ? n <= end : n >= end; n += step) out.push(n); return { out }; } },
-  'list.length': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => ({ out: toArray(inputs.list).length }) },
-  'list.at': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }, { name: 'index', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }, { name: 'index', type: 'number', default: 0 }], evaluate: (inputs) => { const list = toArray(inputs.list); const raw = Math.trunc(inputs.index); const index = raw < 0 ? list.length + raw : raw; return { out: index >= 0 && index < list.length ? list[index] : null }; } },
-  'list.first': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => ({ out: toArray(inputs.list)[0] ?? null }) },
-  'list.last': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => { const list = toArray(inputs.list); return { out: list.length ? list[list.length - 1] : null }; } },
-  'list.map': mapFunctionValueNode('list.map', (list, fn, initial, ctx) => ({ out: list.map((item) => fn.call([item], ctx)) })),
-  'list.filter': mapFunctionValueNode('list.filter', (list, fn, initial, ctx) => ({ out: list.filter((item) => isLoomletTruthy(fn.call([item], ctx))) })),
-  'list.reduce': mapFunctionValueNode('list.reduce', (list, fn, initial, ctx) => ({ out: list.reduce((acc, item) => fn.call([acc, item], ctx), initial) })),
-  'list.join': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }, { name: 'separator', type: 'any', default: ',', kind: 'behavior' }], outputs: [{ name: 'out', type: 'string', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }, { name: 'separator', type: 'any', default: ',' }], evaluate: (inputs) => ({ out: toArray(inputs.list).map((value) => stringifyTextValue(value)).join(stringifyTextValue(inputs.separator)) }) },
-  'list.reverse': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => ({ out: [...toArray(inputs.list)].reverse() }) },
-  'list.sort': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => { const list = [...toArray(inputs.list)]; if (list.every((value) => typeof value === 'number')) list.sort((a, b) => a - b); else list.sort((a, b) => String(a).localeCompare(String(b))); return { out: list }; } },
-  'list.take': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }, { name: 'count', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }, { name: 'count', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: toArray(inputs.list).slice(0, Math.max(0, Math.trunc(inputs.count))) }) },
-  'list.drop': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }, { name: 'count', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }, { name: 'count', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: toArray(inputs.list).slice(Math.max(0, Math.trunc(inputs.count))) }) },
-  'list.concat': { category: 'transform', inputs: Array.from({ length: 4 }, (_, i) => ({ name: `list${i + 1}`, type: 'array', default: undefined, kind: 'behavior' })), outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: Array.from({ length: 4 }, (_, i) => ({ name: `list${i + 1}`, type: 'array', default: undefined })), evaluate: (inputs) => ({ out: collectInputs(inputs, Array.from({ length: 4 }, (_, i) => `list${i + 1}`)).flatMap((value) => toArray(value)) }) },
-
-  'math.add': { category: 'transform', commutative: true, inputs: [{ name: 'a', type: 'number', default: 0, kind: 'behavior' }, { name: 'b', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'a', type: 'number', default: 0 }, { name: 'b', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.a + inputs.b }) },
-  'math.subtract': { category: 'transform', inputs: [{ name: 'a', type: 'number', default: 0, kind: 'behavior' }, { name: 'b', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'a', type: 'number', default: 0 }, { name: 'b', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.a - inputs.b }) },
-  'math.multiply': { category: 'transform', commutative: true, inputs: [{ name: 'a', type: 'number', default: 1, kind: 'behavior' }, { name: 'b', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'a', type: 'number', default: 1 }, { name: 'b', type: 'number', default: 1 }], evaluate: (inputs) => ({ out: inputs.a * inputs.b }) },
-  'math.divide': { category: 'transform', inputs: [{ name: 'a', type: 'number', default: 0, kind: 'behavior' }, { name: 'b', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'a', type: 'number', default: 0 }, { name: 'b', type: 'number', default: 1 }], evaluate: (inputs) => ({ out: inputs.b === 0 ? 0 : inputs.a / inputs.b }) },
-  'math.mod': { category: 'transform', inputs: [{ name: 'a', type: 'number', default: 0, kind: 'behavior' }, { name: 'b', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'a', type: 'number', default: 0 }, { name: 'b', type: 'number', default: 1 }], evaluate: (inputs) => ({ out: inputs.b === 0 ? 0 : ((inputs.a % inputs.b) + inputs.b) % inputs.b }) },
-  'math.abs': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.abs(inputs.value) }) },
-  'math.clamp': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'min', type: 'number', default: 0, kind: 'behavior' }, { name: 'max', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'min', type: 'number', default: 0 }, { name: 'max', type: 'number', default: 1 }], evaluate: (inputs) => ({ out: inputs.min > inputs.max ? inputs.min : Math.max(inputs.min, Math.min(inputs.max, inputs.value)) }) },
-  'math.map': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'inMin', type: 'number', default: 0, kind: 'behavior' }, { name: 'inMax', type: 'number', default: 1, kind: 'behavior' }, { name: 'outMin', type: 'number', default: 0, kind: 'behavior' }, { name: 'outMax', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'inMin', type: 'number', default: 0 }, { name: 'inMax', type: 'number', default: 1 }, { name: 'outMin', type: 'number', default: 0 }, { name: 'outMax', type: 'number', default: 1 }, { name: 'clamp', type: 'boolean', default: false }], evaluate: (inputs, params) => { if (inputs.inMax === inputs.inMin) return { out: inputs.outMin }; let t = (inputs.value - inputs.inMin) / (inputs.inMax - inputs.inMin); if (params.clamp === true) t = Math.max(0, Math.min(1, t)); return { out: inputs.outMin + (inputs.outMax - inputs.outMin) * t }; } },
-  'math.lerp': { category: 'transform', inputs: [{ name: 'a', type: 'number', default: 0, kind: 'behavior' }, { name: 'b', type: 'number', default: 1, kind: 'behavior' }, { name: 't', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'a', type: 'number', default: 0 }, { name: 'b', type: 'number', default: 1 }, { name: 't', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.a + (inputs.b - inputs.a) * inputs.t }) },
-  'math.smoothstep': { category: 'transform', inputs: [{ name: 'x', type: 'number', default: 0, kind: 'behavior' }, { name: 'edge0', type: 'number', default: 0, kind: 'behavior' }, { name: 'edge1', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'x', type: 'number', default: 0 }, { name: 'edge0', type: 'number', default: 0 }, { name: 'edge1', type: 'number', default: 1 }], evaluate: (inputs) => { if (inputs.edge0 === inputs.edge1) return { out: inputs.x < inputs.edge0 ? 0 : 1 }; let t = (inputs.x - inputs.edge0) / (inputs.edge1 - inputs.edge0); t = Math.max(0, Math.min(1, t)); return { out: t * t * (3 - 2 * t) }; } },
-  'math.cosine': { category: 'transform', inputs: [{ name: 't', type: 'number', default: 0, kind: 'behavior' }, { name: 'freq', type: 'number', default: 1, kind: 'behavior' }, { name: 'amplitude', type: 'number', default: 1, kind: 'behavior' }, { name: 'phase', type: 'number', default: 0, kind: 'behavior' }, { name: 'offset', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'freq', type: 'number', default: 1 }, { name: 'amplitude', type: 'number', default: 1 }, { name: 'phase', type: 'number', default: 0 }, { name: 'offset', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.cos(inputs.t * inputs.freq * 2 * Math.PI + inputs.phase) * inputs.amplitude + inputs.offset }) },
-  'math.floor': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.floor(inputs.value) }) },
-  'math.ceil': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.ceil(inputs.value) }) },
-  'math.round': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.round(inputs.value) }) },
-  'math.min': { category: 'transform', commutative: true, inputs: [{ name: 'a', type: 'number', default: 0, kind: 'behavior' }, { name: 'b', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'a', type: 'number', default: 0 }, { name: 'b', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.min(inputs.a, inputs.b) }) },
-  'math.max': { category: 'transform', commutative: true, inputs: [{ name: 'a', type: 'number', default: 0, kind: 'behavior' }, { name: 'b', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'a', type: 'number', default: 0 }, { name: 'b', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.max(inputs.a, inputs.b) }) },
-  'math.tan': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.tan(inputs.value) }) },
-  'math.sqrt': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: Math.sqrt(inputs.value) }) },
-  'math.pow': { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'exponent', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'exponent', type: 'number', default: 1 }], evaluate: (inputs) => ({ out: Math.pow(inputs.value, inputs.exponent) }) },
-
-  'random.value': { category: 'source', inputs: [], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [], evaluate: () => ({ out: Math.random() }) },
-  'random.range': { category: 'transform', inputs: [{ name: 'min', type: 'number', default: 0, kind: 'behavior' }, { name: 'max', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'min', type: 'number', default: 0 }, { name: 'max', type: 'number', default: 1 }], evaluate: (inputs) => ({ out: inputs.min + Math.random() * (inputs.max - inputs.min) }) },
-  'random.int': { category: 'transform', inputs: [{ name: 'min', type: 'number', default: 0, kind: 'behavior' }, { name: 'max', type: 'number', default: 1, kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'min', type: 'number', default: 0 }, { name: 'max', type: 'number', default: 1 }], evaluate: (inputs) => { const min = Math.ceil(Math.min(inputs.min, inputs.max)); const max = Math.floor(Math.max(inputs.min, inputs.max)); return { out: Math.floor(Math.random() * (max - min + 1)) + min }; } },
-  'random.choice': { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => { const list = toArray(inputs.list); return { out: list.length ? list[Math.floor(Math.random() * list.length)] : null }; } },
-
-  'debug.inspect': { category: 'transform', inputs: [{ name: 'value', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'string', kind: 'behavior' }], params: [{ name: 'value', type: 'any', default: null }], evaluate: (inputs) => ({ out: inspectValue(inputs.value) }) },
-  'debug.trace': { category: 'transform', inputs: [{ name: 'value', type: 'any', default: null, kind: 'behavior' }, { name: 'label', type: 'string', default: 'trace', kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'value', type: 'any', default: null }, { name: 'label', type: 'string', default: 'trace' }], evaluate: (inputs, params, ctx) => { ctx.engine?._recordEffect({ type: 'debug.trace', label: inputs.label, value: inputs.value, nodeId: ctx.currentNodeId }); return { out: inputs.value }; } },
-  'debug.assert': { category: 'transform', inputs: [{ name: 'condition', type: 'any', default: false, kind: 'behavior' }, { name: 'message', type: 'string', default: 'Assertion failed', kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'condition', type: 'any', default: false }, { name: 'message', type: 'string', default: 'Assertion failed' }], evaluate: (inputs) => { if (!inputs.condition) throw new LoomError('ASSERTION_FAILED', stringifyTextValue(inputs.message) || 'Assertion failed'); return { out: true }; } },
-
-  'fs.readText': { category: 'source', inputs: [{ name: 'path', type: 'string', default: '', kind: 'behavior' }], outputs: [{ name: 'out', type: 'string', kind: 'behavior' }], params: [{ name: 'path', type: 'string', default: '' }], evaluate: (inputs) => ({ out: getNodeFs().readFileSync(String(inputs.path), 'utf8') }) },
-  'fs.writeText': { category: 'sink', inputs: [{ name: 'path', type: 'string', default: '', kind: 'behavior' }, { name: 'value', type: 'any', default: '', kind: 'behavior' }], outputs: [], params: [{ name: 'path', type: 'string', default: '' }, { name: 'value', type: 'any', default: '' }], evaluate: (inputs) => { const fs = getNodeFs(); const path = getNodePath(); fs.mkdirSync(path.dirname(String(inputs.path)), { recursive: true }); fs.writeFileSync(String(inputs.path), stringifyTextValue(inputs.value), 'utf8'); return {}; } },
-  'fs.exists': { category: 'source', inputs: [{ name: 'path', type: 'string', default: '', kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'path', type: 'string', default: '' }], evaluate: (inputs) => ({ out: getNodeFs().existsSync(String(inputs.path)) }) },
-  'fs.list': { category: 'source', inputs: [{ name: 'path', type: 'string', default: '.', kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'path', type: 'string', default: '.' }], evaluate: (inputs) => ({ out: getNodeFs().readdirSync(String(inputs.path)) }) },
-
-  // Scene Sync effect nodes
-  'scene.setPosition': {
-    category: 'sink',
-    inputs: [
-      { name: 'objectId', type: 'string', default: '', kind: 'behavior' },
-      { name: 'x', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'y', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'z', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'objectId', type: 'string', default: '' },
-      { name: 'x', type: 'number', default: 0 },
-      { name: 'y', type: 'number', default: 0 },
-      { name: 'z', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      ctx.engine?._recordEffect({
-        type: 'scene.setPosition',
-        objectId: inputs.objectId,
-        position: [inputs.x, inputs.y, inputs.z],
-        target: 'scenesync',
-        nodeId: ctx.currentNodeId
-      });
-      return {};
-    }
-  },
-
-  'scene.setRotation': {
-    category: 'sink',
-    inputs: [
-      { name: 'objectId', type: 'string', default: '', kind: 'behavior' },
-      { name: 'x', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'y', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'z', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'w', type: 'number', default: 1, kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'objectId', type: 'string', default: '' },
-      { name: 'x', type: 'number', default: 0 },
-      { name: 'y', type: 'number', default: 0 },
-      { name: 'z', type: 'number', default: 0 },
-      { name: 'w', type: 'number', default: 1 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      ctx.engine?._recordEffect({
-        type: 'scene.setRotation',
-        objectId: inputs.objectId,
-        rotation: [inputs.x, inputs.y, inputs.z, inputs.w],
-        target: 'scenesync',
-        nodeId: ctx.currentNodeId
-      });
-      return {};
-    }
-  },
-
-  'scene.setScale': {
-    category: 'sink',
-    inputs: [
-      { name: 'objectId', type: 'string', default: '', kind: 'behavior' },
-      { name: 'x', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'y', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'z', type: 'number', default: 1, kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'objectId', type: 'string', default: '' },
-      { name: 'x', type: 'number', default: 1 },
-      { name: 'y', type: 'number', default: 1 },
-      { name: 'z', type: 'number', default: 1 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      ctx.engine?._recordEffect({
-        type: 'scene.setScale',
-        objectId: inputs.objectId,
-        scale: [inputs.x, inputs.y, inputs.z],
-        target: 'scenesync',
-        nodeId: ctx.currentNodeId
-      });
-      return {};
-    }
-  },
-
-  'scene.offsetPosition': {
-    category: 'sink',
-    inputs: [
-      { name: 'objectId', type: 'string', default: '', kind: 'behavior' },
-      { name: 'x', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'y', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'z', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'objectId', type: 'string', default: '' },
-      { name: 'x', type: 'number', default: 0 },
-      { name: 'y', type: 'number', default: 0 },
-      { name: 'z', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      ctx.engine?._recordEffect({
-        type: 'scene.offsetPosition',
-        objectId: inputs.objectId,
-        offset: [inputs.x, inputs.y, inputs.z],
-        target: 'scenesync',
-        nodeId: ctx.currentNodeId
-      });
-      return {};
-    }
-  },
-
-  // DOM シンクノード
-  setText: {
-    category: 'sink',
-    inputs: [
-      { name: 'value', type: 'any', default: '', kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'target', type: 'string', default: '' }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      if (!params.target) return {};
-      const el = document.querySelector(params.target);
-      if (el) el.textContent = String(inputs.value);
-      return {};
-    }
-  },
-
-  setStyle: {
-    category: 'sink',
-    inputs: [
-      { name: 'value', type: 'any', default: '', kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'target', type: 'string', default: '' },
-      { name: 'property', type: 'string', default: '' },
-      { name: 'unit', type: 'string', default: '' }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      if (!params.target || !params.property) return {};
-      const el = document.querySelector(params.target);
-      if (el) el.style[params.property] = String(inputs.value) + params.unit;
-      return {};
-    }
-  },
-  setClass: {
-    category: 'sink',
-    inputs: [
-      { name: 'enabled', type: 'boolean', default: true, kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'target', type: 'string', default: '' },
-      { name: 'className', type: 'string', default: '' }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      if (!params.target || !params.className) return {};
-      const el = document.querySelector(params.target);
-      if (!el) return {};
-      el.classList.toggle(params.className, Boolean(inputs.enabled));
-      return {};
-    }
-  },
-  setCssVar: {
-    category: 'sink',
-    inputs: [
-      { name: 'value', type: 'any', default: 0, kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'target', type: 'string', default: '' },
-      { name: 'name', type: 'string', default: '' },
-      { name: 'unit', type: 'string', default: '' }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      if (!params.target || !params.name) return {};
-      if (inputs.value === null || inputs.value === undefined) return {};
-      const el = document.querySelector(params.target);
-      if (!el) return {};
-      const cssVarName = params.name.startsWith('--') ? params.name : `--${params.name}`;
-      el.style.setProperty(cssVarName, String(inputs.value) + params.unit);
-      return {};
-    }
-  },
-  setTransform2D: {
-    category: 'sink',
-    inputs: [
-      { name: 'x', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'y', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'scale', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'rotate', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'target', type: 'string', default: '' },
-      { name: 'unit', type: 'string', default: 'px' },
-      { name: 'rotateUnit', type: 'string', default: 'deg' }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      if (!params.target) return {};
-      const el = document.querySelector(params.target);
-      if (!el) return {};
-      el.style.transform = `translate(${inputs.x}${params.unit}, ${inputs.y}${params.unit}) scale(${inputs.scale}) rotate(${inputs.rotate}${params.rotateUnit})`;
-      return {};
-    }
-  },
-
-  setAttr: {
-    category: 'sink',
-    inputs: [
-      { name: 'value', type: 'any', default: '', kind: 'behavior' }
-    ],
-    outputs: [],
-    params: [
-      { name: 'target', type: 'string', default: '' },
-      { name: 'name', type: 'string', default: '' }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      if (!params.target || !params.name) return {};
-      const el = document.querySelector(params.target);
-      if (el) el.setAttribute(params.name, String(inputs.value));
-      return {};
-    }
-  },
-
-  'time.serverClock': {
-    category: 'source',
-    inputs: [],
-    outputs: [{ name: 't', type: 'number', kind: 'behavior' }],
-    params: [],
-    evaluate: (inputs, params, ctx) => ({ t: ctx.time })
-  },
-
-  'math.sine': {
-    category: 'transform',
-    inputs: [
-      { name: 't', type: 'number', default: 0, kind: 'behavior' },
-      { name: 'freq', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'amplitude', type: 'number', default: 1, kind: 'behavior' },
-      { name: 'offset', type: 'number', default: 0, kind: 'behavior' }
-    ],
-    outputs: [{ name: 'out', type: 'number', kind: 'behavior' }],
-    params: [
-      { name: 'freq', type: 'number', default: 1 },
-      { name: 'amplitude', type: 'number', default: 1 },
-      { name: 'offset', type: 'number', default: 0 }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const t = inputs.t;
-      const freq = inputs.freq;
-      const amplitude = inputs.amplitude;
-      const offset = inputs.offset;
-      return { out: Math.sin(t * freq * 2 * Math.PI) * amplitude + offset };
-    }
-  },
-
-  log: {
-    category: 'output',
-    inputs: [
-      { name: 'value', type: 'any', default: undefined, kind: 'behavior' }
-    ],
-    outputs: [
-      { name: 'value', type: 'any', kind: 'behavior' }
-    ],
-    params: [
-      { name: 'label', type: 'string', default: '' }
-    ],
-    evaluate: (inputs, params, ctx) => {
-      const label = params.label ?? '';
-      const message = label ? `${label}: ${inspectValue(inputs.value)}` : inspectValue(inputs.value);
-      ctx.engine?._recordEffect({ type: 'log', message, nodeId: ctx.currentNodeId });
-      return { value: inputs.value };
-    }
-  }
-};
 
 export class Loom {
   constructor(graph) {
@@ -2233,7 +1166,9 @@ export class Loom {
 }
 
 export function createDefaultNodeRegistry() {
-  return createNodeRegistry(BUILTIN_NODE_TYPES);
+  const registry = createNodeRegistry();
+  registerBuiltinNodes(registry);
+  return registry;
 }
 
 export const DEFAULT_NODE_REGISTRY = createDefaultNodeRegistry();
