+++ src/nodes/list.js
@@ -1,0 +1,19 @@
+import { collectInputs, isLoomletTruthy, mapFunctionValueNode, stringifyTextValue, toArray } from './helpers.js';
+
+export function registerListNodes(registry) {
+  registry.registerNodeType('list.at', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }, { name: 'index', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }, { name: 'index', type: 'number', default: 0 }], evaluate: (inputs) => { const list = toArray(inputs.list); const raw = Math.trunc(inputs.index); const index = raw < 0 ? list.length + raw : raw; return { out: index >= 0 && index < list.length ? list[index] : null }; } });
+  registry.registerNodeType('list.concat', { category: 'transform', inputs: Array.from({ length: 4 }, (_, i) => ({ name: `list${i + 1}`, type: 'array', default: undefined, kind: 'behavior' })), outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: Array.from({ length: 4 }, (_, i) => ({ name: `list${i + 1}`, type: 'array', default: undefined })), evaluate: (inputs) => ({ out: collectInputs(inputs, Array.from({ length: 4 }, (_, i) => `list${i + 1}`)).flatMap((value) => toArray(value)) }) });
+  registry.registerNodeType('list.drop', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }, { name: 'count', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }, { name: 'count', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: toArray(inputs.list).slice(Math.max(0, Math.trunc(inputs.count))) }) });
+  registry.registerNodeType('list.filter', mapFunctionValueNode('list.filter', (list, fn, initial, ctx) => ({ out: list.filter((item) => isLoomletTruthy(fn.call([item], ctx))) })),);
+  registry.registerNodeType('list.first', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => ({ out: toArray(inputs.list)[0] ?? null }) });
+  registry.registerNodeType('list.join', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }, { name: 'separator', type: 'any', default: ',', kind: 'behavior' }], outputs: [{ name: 'out', type: 'string', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }, { name: 'separator', type: 'any', default: ',' }], evaluate: (inputs) => ({ out: toArray(inputs.list).map((value) => stringifyTextValue(value)).join(stringifyTextValue(inputs.separator)) }) });
+  registry.registerNodeType('list.last', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => { const list = toArray(inputs.list); return { out: list.length ? list[list.length - 1] : null }; } });
+  registry.registerNodeType('list.length', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'number', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => ({ out: toArray(inputs.list).length }) });
+  registry.registerNodeType('list.map', mapFunctionValueNode('list.map', (list, fn, initial, ctx) => ({ out: list.map((item) => fn.call([item], ctx)) })),);
+  registry.registerNodeType('list.of', { category: 'transform', inputs: Array.from({ length: 8 }, (_, i) => ({ name: `value${i + 1}`, type: 'any', default: undefined, kind: 'behavior' })), outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: Array.from({ length: 8 }, (_, i) => ({ name: `value${i + 1}`, type: 'any', default: undefined })), evaluate: (inputs) => ({ out: collectInputs(inputs, Array.from({ length: 8 }, (_, i) => `value${i + 1}`)) }) });
+  registry.registerNodeType('list.range', { category: 'transform', inputs: [{ name: 'start', type: 'number', default: 0, kind: 'behavior' }, { name: 'end', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'start', type: 'number', default: 0 }, { name: 'end', type: 'number', default: 0 }], evaluate: (inputs) => { const start = Math.trunc(inputs.start); const end = Math.trunc(inputs.end); const step = start <= end ? 1 : -1; const out = []; for (let n = start; step > 0 ? n <= end : n >= end; n += step) out.push(n); return { out }; } });
+  registry.registerNodeType('list.reduce', mapFunctionValueNode('list.reduce', (list, fn, initial, ctx) => ({ out: list.reduce((acc, item) => fn.call([acc, item], ctx), initial) })),);
+  registry.registerNodeType('list.reverse', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => ({ out: [...toArray(inputs.list)].reverse() }) });
+  registry.registerNodeType('list.sort', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }], evaluate: (inputs) => { const list = [...toArray(inputs.list)]; if (list.every((value) => typeof value === 'number')) list.sort((a, b) => a - b); else list.sort((a, b) => String(a).localeCompare(String(b))); return { out: list }; } });
+  registry.registerNodeType('list.take', { category: 'transform', inputs: [{ name: 'list', type: 'array', default: [], kind: 'behavior' }, { name: 'count', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'array', kind: 'behavior' }], params: [{ name: 'list', type: 'array', default: [] }, { name: 'count', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: toArray(inputs.list).slice(0, Math.max(0, Math.trunc(inputs.count))) }) });
+}
