+++ src/nodes/logic.js
@@ -1,0 +1,13 @@
+export function registerLogicNodes(registry) {
+  registry.registerNodeType('logic.and', { category: 'transform', commutative: true, inputs: [{ name: 'a', type: 'any', default: false, kind: 'behavior' }, { name: 'b', type: 'any', default: false, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'a', type: 'any', default: false }, { name: 'b', type: 'any', default: false }], evaluate: (inputs) => ({ out: Boolean(inputs.a && inputs.b) }) });
+  registry.registerNodeType('logic.equals', { category: 'transform', inputs: [{ name: 'value', type: 'any', default: null, kind: 'behavior' }, { name: 'other', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'any', default: null }, { name: 'other', type: 'any', default: null }], evaluate: (inputs) => ({ out: Object.is(inputs.value, inputs.other) }) });
+  registry.registerNodeType('logic.greaterOrEqual', { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'other', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'other', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.value >= inputs.other }) });
+  registry.registerNodeType('logic.greaterThan', { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'other', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'other', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.value > inputs.other }) });
+  registry.registerNodeType('logic.lessOrEqual', { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'other', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'other', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.value <= inputs.other }) });
+  registry.registerNodeType('logic.lessThan', { category: 'transform', inputs: [{ name: 'value', type: 'number', default: 0, kind: 'behavior' }, { name: 'other', type: 'number', default: 0, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'number', default: 0 }, { name: 'other', type: 'number', default: 0 }], evaluate: (inputs) => ({ out: inputs.value < inputs.other }) });
+  registry.registerNodeType('logic.not', { category: 'transform', inputs: [{ name: 'value', type: 'any', default: false, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'any', default: false }], evaluate: (inputs) => ({ out: !inputs.value }) });
+  registry.registerNodeType('logic.notEquals', { category: 'transform', inputs: [{ name: 'value', type: 'any', default: null, kind: 'behavior' }, { name: 'other', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'value', type: 'any', default: null }, { name: 'other', type: 'any', default: null }], evaluate: (inputs) => ({ out: !Object.is(inputs.value, inputs.other) }) });
+  registry.registerNodeType('logic.or', { category: 'transform', commutative: true, inputs: [{ name: 'a', type: 'any', default: false, kind: 'behavior' }, { name: 'b', type: 'any', default: false, kind: 'behavior' }], outputs: [{ name: 'out', type: 'boolean', kind: 'behavior' }], params: [{ name: 'a', type: 'any', default: false }, { name: 'b', type: 'any', default: false }], evaluate: (inputs) => ({ out: Boolean(inputs.a || inputs.b) }) });
+  registry.registerNodeType('logic.select', { category: 'transform', inputs: [{ name: 'condition', type: 'any', default: false, kind: 'behavior' }, { name: 'whenTrue', type: 'any', default: null, kind: 'behavior' }, { name: 'whenFalse', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'condition', type: 'any', default: false }, { name: 'whenTrue', type: 'any', default: null }, { name: 'whenFalse', type: 'any', default: null }], evaluate: (inputs) => ({ out: inputs.condition ? inputs.whenTrue : inputs.whenFalse }) });
+  registry.registerNodeType('logic.when', { category: 'transform', inputs: [{ name: 'condition', type: 'any', default: false, kind: 'behavior' }, { name: 'value', type: 'any', default: null, kind: 'behavior' }], outputs: [{ name: 'out', type: 'any', kind: 'behavior' }], params: [{ name: 'condition', type: 'any', default: false }, { name: 'value', type: 'any', default: null }], evaluate: (inputs) => ({ out: inputs.condition ? inputs.value : null }) });
+}
